[#9]: Refactoring the scripts. all the python code should be placed under the library root. Also, let's just call the scripts "main". Moving un-main scripts to identify_idioms/examples.
