+++ identify_idioms/main/scrape_slide_idiom_alts.py
@@ -1,10 +1,10 @@
 import json
-from requests import HTTPError
-from identify_idioms.builders import IdiomAltsBuilder
-from identify_idioms.loaders import SlideIdiomsLoader
-from identify_idioms.config import SLIDE_TSV, SLIDE_IDIOM_ALTS_TSV
 import csv
 from os import path
+from requests import HTTPError
+from identify_idioms.builders import IdiomAltsBuilder
+from identify_idioms.loaders import load_slide_idioms
+from identify_idioms.paths import SLIDE_IDIOM_ALTS_TSV
 
 
 def main():
@@ -16,7 +16,7 @@ def main():
         tsv_writer = csv.writer(fh, delimiter="\t")
         schema = ['slide_idiom', 'alts']
         tsv_writer.writerow(schema)
-        idioms = SlideIdiomsLoader(SLIDE_TSV).load()
+        idioms = load_slide_idioms()
         for idx, idiom in enumerate(idioms):
             try:
                 alts = alts_builder.construct(idiom)
